pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "plasmid-priority"
version = "0.3.0"
description = "Retrospective genomic surveillance prioritization for operational plasmid backbone classes."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "duckdb>=1.5.1",
  "numba>=0.60",
  "numpy>=2.0",
  "pandas>=2.2",
  "pandera>=0.20",
  "pydantic>=2.7",
  "pydantic-settings>=2.3",
  "pyyaml>=6",
  "scikit-learn>=1.5",
  "scipy>=1.13",
  "threadpoolctl>=3.0",
  "xlrd>=2.0.1",
]

[project.optional-dependencies]
analysis = [
  "interpret>=0.7.8",
  "matplotlib>=3.10",
  "networkx>=3.0",
  "pillow>=12.2.0",
  "pyarrow>=16",
  "shap>=0.44",
  "snakemake>=8",
]
# Phase 1 SOTA dependencies
sota = [
  "lifelines>=0.30",
  "scikit-survival>=0.23",
  "igraph>=0.11",
  "leidenalg>=0.10",
  "torch>=2.3",
  "biopython>=1.84",
]
dev = [
  "pytest>=9.0.3",
  "pytest-cov>=5",
  "pygments>=2.20.0",
  "mypy>=1.10",
  "ruff>=0.4",
  "pip-audit>=0.3",
  "pre-commit>=3.7",
  "commitizen>=3.0",
  "mkdocs>=1.6",
  "mkdocs-material>=9.5",
  "mkdocstrings[python]>=0.25",
  "types-PyYAML>=6.0.12",
  "pandas-stubs>=2.2",
  "hypothesis>=6.100",
  "mutmut>=2.5.1",
]
api = [
  "fastapi>=0.111",
  "uvicorn[standard]>=0.30",
]
engineering = [
  "prefect>=3.0",
  "dvc>=3.55",
]
tree-models = [
  "lightgbm>=4.0",
  "xgboost>=2.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_ignores = true
ignore_missing_imports = true
warn_redundant_casts = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "plasmid_priority.validation.metrics",
    "plasmid_priority.validation.vif",
    "plasmid_priority.validation.boundaries",
]
strict = true
disallow_untyped_defs = true
no_implicit_reexport = true

# Suppress pandas type-narrowing false positives (unreachable warnings after isinstance checks)
[[tool.mypy.overrides]]
module = [
    "plasmid_priority.labels.probabilistic",
    "plasmid_priority.backbone.graph_clustering",
]
warn_unreachable = false

[[tool.mypy.overrides]]
module = [
    "plasmid_priority.pipeline.branches",
    "plasmid_priority.reporting.build_reports_script_impl",
]
ignore_errors = true

# Scripts and Test files — relax strict checks
[[tool.mypy.overrides]]
module = [
]
disallow_untyped_defs = false
warn_return_any = false
ignore_errors = true

# Optional-dependency src modules (torch, prefect, igraph, biopython)
[[tool.mypy.overrides]]
module = [
    "plasmid_priority.labels.coteaching",
    "plasmid_priority.modeling.multi_task",
    "plasmid_priority.pipeline.prefect_flow",
    "plasmid_priority.backbone.graph_clustering",
    "plasmid_priority.embedding.kmer_baseline",
]
disallow_untyped_defs = false
warn_return_any = false
ignore_errors = false

[tool.pytest.ini_options]
addopts = "-x -q --tb=short"
testpaths = ["tests"]
markers = [
  "slow: marks tests as slow (deselect with '-m not slow')",
  "integration: marks integration tests requiring data",
]
filterwarnings = [
  'ignore:Falling back to stratified_repeated because resolved_year is missing for temporal_group\.:UserWarning',
  'ignore:\d+ sample\(s\) never appeared in any test fold:UserWarning',
  'ignore:datetime\.datetime\.utcnow\(\) is deprecated.*:DeprecationWarning',
  'ignore:Starting with pandas version 4\.0 all arguments of (var|mean|std) will be keyword-only\.:Warning',
]

[tool.coverage.run]
source = ["src/plasmid_priority"]
omit = [
  "*/__pycache__/*",
  "*/tests/*",
  "src/plasmid_priority/api/*",
  "src/plasmid_priority/annotate/*",
  "src/plasmid_priority/*/cli.py",
  "src/plasmid_priority/reporting/*_impl.py",
  "src/plasmid_priority/shared/branch_base.py",
  "src/plasmid_priority/pipeline/branches.py",
  "src/plasmid_priority/qc/input_checks.py",
  "src/plasmid_priority/features/host.py",
  "src/plasmid_priority/features/host_range.py",
  "src/plasmid_priority/features/interaction.py",
  "src/plasmid_priority/features/core_impl.py",
  "src/plasmid_priority/features/transfer.py",
  "src/plasmid_priority/features/transmission.py",
  "src/plasmid_priority/modeling/firthlogist_bridge.py",
  "src/plasmid_priority/modeling/fit_config.py",
  "src/plasmid_priority/modeling/module_a_impl.py",
  "src/plasmid_priority/modeling/nested_cv.py",
  "src/plasmid_priority/modeling/task.py",
  "src/plasmid_priority/modeling/tree_models.py",
  "src/plasmid_priority/utils/parquet_utils.py",
  "src/plasmid_priority/utils/threshold_sensitivity.py",
  "src/plasmid_priority/survival/*.py",
  "src/plasmid_priority/embedding/*.py",
  "src/plasmid_priority/labels/coteaching.py",
  "src/plasmid_priority/labels/counterfactual.py",
  "src/plasmid_priority/backbone/graph_clustering.py",
  "src/plasmid_priority/modeling/ft_transformer.py",
  "src/plasmid_priority/modeling/multi_task.py",
  "src/plasmid_priority/modeling/conformal.py",
  "src/plasmid_priority/modeling/deep_ensemble.py",
  "src/plasmid_priority/modeling/nested_cv_bayesian.py",
  "src/plasmid_priority/modeling/attention_consensus.py",
  "src/plasmid_priority/modeling/ood_detection.py",
  "src/plasmid_priority/modeling/torch_utils.py",
  "src/plasmid_priority/consensus/optuna_weights.py",
  "src/plasmid_priority/reporting/figures.py",
]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
    "@overload",
    "def __repr__",
    "if __name__ == .__main__.:",
]
fail_under = 70  # Faz 4 hedef: %70 coverage
precision = 2
show_missing = true

[tool.coverage.html]
directory = "htmlcov"

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

[tool.ruff.lint.per-file-ignores]
"scripts/*.py" = ["E402", "E501"]
"tests/*.py" = ["E402", "E501"]
"src/plasmid_priority/reporting/enrichment.py" = ["E402", "E501"]
"src/plasmid_priority/reporting/model_audit.py" = ["E501"]
"src/plasmid_priority/reporting/model_audit_impl.py" = ["E501"]
"src/plasmid_priority/reporting/build_reports_script_impl.py" = ["E402", "E501", "I001"]
"src/plasmid_priority/modeling/module_a_impl.py" = ["E501"]
"src/plasmid_priority/reporting/figures_impl.py" = ["E501"]
"src/plasmid_priority/reporting/advanced_audits_impl.py" = ["E501"]
"src/plasmid_priority/reporting/candidate_tables.py" = ["E501"]
"src/plasmid_priority/features/core_impl.py" = ["E501"]

[tool.commitizen]
name = "cz_conventional_commits"
tag_format = "v$version"
version_scheme = "semver"
version_provider = "pep621"
update_changelog_on_bump = true

[tool.mutmut]
paths_to_mutate = ["src/plasmid_priority"]
tests_dir = "tests"
runner = "PYTHONPATH=src pytest -q -o addopts=''"
